+++ src/ios/CDVInAppBrowserOptions.h
@@ -37,7 +37,9 @@
 
 @property (nonatomic, copy) NSString* presentationstyle;
 @property (nonatomic, copy) NSString* transitionstyle;
+@property (nonatomic, copy) NSString* statusbarstyle;
 
+@property (nonatomic, assign) BOOL statusbartransparent;
 @property (nonatomic, assign) BOOL enableviewportscale;
 @property (nonatomic, assign) BOOL mediaplaybackrequiresuseraction;
 @property (nonatomic, assign) BOOL allowinlinemediaplayback;