+++ src/ios/CDVInAppBrowserOptions.m
@@ -45,6 +45,8 @@
         self.toolbarcolor = nil;
         self.toolbartranslucent = YES;
         self.beforeload = @"";
+        self.statusbarstyle = @"dark";
+        self.statusbartransparent = NO;
     }
 
     return self;